refactor: moved Toolbar to modules
